ciaaTest project modified for DAC testing.

+++ workspace_lpcopen/libCIAA/inc/ciaaAOUT.h
@@ -11,6 +11,6 @@
 #include "chip.h"
 
 void ciaaAOUTInit(void);
-void ciaaAOUTSet(uint8_t level);
+void ciaaAOUTSet(float level);
 
 #endif /* CIAAAOUT_H_ */
