+++ workspace_lpcopen/libCIAA/src/ciaaAOUT.c
@@ -18,9 +18,9 @@ void ciaaAOUTInit(void)
 }
 
 /* [in] level: Output level (0 - 100%) */
-void ciaaAOUTSet(uint8_t level)
+void ciaaAOUTSet(float level)
 {
 	if(level > 100) level = 100;
 
-	Chip_DAC_UpdateValue(LPC_DAC, (level * 0x3FF) / 100);
+	Chip_DAC_UpdateValue(LPC_DAC, (uint32_t)((level * (float)0x3FF) / 100.0));
 }
